+++ src/components/Slide/Slide.style.js
@@ -2,6 +2,21 @@ import styled from 'styled-components';
 import { FONT_FAMILY } from 'constants/typography';
 import { WHITE } from 'constants/colours';
 
+export const Error = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: ${WHITE};
+  display: flex;
+  flex: 1;
+  justify-content: center;
+  align-items: center;
+  font-family: ${FONT_FAMILY};
+  font-size: 24px;
+`;
+
 export const Image = styled.img`
   width: ${({ props }) => props.width || 'auto'};
   height: ${({ props }) => props.height || 'auto'};