+++ src/server.js
@@ -3,22 +3,35 @@ import express from 'express';
 import { compile } from 'ejs';
 import { readFileSync, readJson } from 'fs-extra';
 import { join } from 'path';
+import fetch from 'node-fetch';
 
 const port = process.env.WEPBACK_PORT || 3005;
 const app = express();
 
-const str = readFileSync(join(__dirname, 'index.html'), { encoding: 'utf8' });
-const template = compile(str);
-
 app.use(express.static(join(__dirname, 'public')));
 
+// Route all media requests through pexels.com for dummy images
+app.get('/media/*', (req, res) => {
+  let url = req.originalUrl.replace('/media/', '');
+  url = `https://images.pexels.com/photos/${url}`;
+
+  return fetch(url)
+    .then(response => response.buffer())
+    .then(buffer => res.send(buffer));
+});
+
+// For all other requests, show the index page
 app.get('*', (req, res) => {
+  const str = readFileSync(join(__dirname, 'index.html'), { encoding: 'utf8' });
+  const template = compile(str);
+
   readJson(join(__dirname, 'public/scripts/manifest.json')).then((manifest) => {
     const html = template({
       scripts: {
         manifest: `/scripts/${manifest['manifest.js']}`,
         vendor: `/scripts/${manifest['vendor.js']}`,
         photine: `/scripts/${manifest['photine.js']}`,
+        serviceWorker: `/scripts/${manifest['serviceWorker.js']}`,
       },
     });